chore(deps): bump tauri, windows, uiautomation
Consolidate Dependabot PRs #24-26. Adapt CallNextHookEx to windows 0.61 Option<HHOOK> API and uiautomation get_process_id u32 return type.

+++ src-tauri/src/observe/reader_win.rs
@@ -10,7 +10,7 @@ thread_local! {
 }
 
 fn focused_in_own_process(element: &UIElement) -> bool {
-    let own_pid = std::process::id() as i32;
+    let own_pid = std::process::id();
     let mut current = element.clone();
     for _ in 0..8 {
         if current.get_process_id().ok() == Some(own_pid) {
